// src/app/classes/interfaces.ts
export interface Commerce {
  usuario?: Usuario[];
  fecha?: string;
  modified?: string;
  coords?: string;
  bio?: string;
  cellphone?: string;
  street?: string;

  status?: string;

  telephone?: string;
  whatsapp?: string;
  delivery?: string;
  card?: string;
  opening_hours?: string;
  closing_schedule?: string;

  province?: string;
  municipality?: string;
  location?: string;
  neighborhood?: string;
  rubro?: string;
  commerce?: string;
  img?: string;
  _id?: string;
  visitas?: number;
  clickswhats?: number;
  clickscell?: number;
  clickstel?: number;
  shares?: number;
  favs?: number;

  linkinsta?: string;
  linkface?: string;
  linkweb?: string;

  home?: string;

  bussines_days?: string;
  premium?: string;
  coupons?: string;

  img_2?: string;
  img_3?: string;
  img_4?: string;
  img_5?: string;

  tags?: string;
  notes?: string;

  vendedor?: string;
  plan?: string;
  prioridad?: number;
}
// ...
export interface Usuario {
  avatar?: string;
  _id?: string;
  nombre?: string;
  email?: string;
  password?: string;
  password_c?: string;
  role?: string;
  fb?: string;
  fecha?: string;

  interests?: string;
  version?: string;
  notification_id?: string;
}